pyproject.toml
[tool.poetry]
name = "preprocessing"
version = "0.0.1"
description = "This module contains the code for preprocessing and postprocessing configuration"
authors = ["TataiQ <f.>"]
repository = "https://github.com/./PreProcessing"
documentation = "https://..github.io/PreProcessing/"
readme = "README.md"
packages = [
  {include = "preprocessing"}
]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
opencv-python = "^4.8.1.78"
numpy = "^1.26.1"
pandas = "^2.1.1"
kafka-python = "^2.0.2"
scipy = "^1.11.3"
redis = "^5.0.1"
shared-memory-dict = "^0.7.2"
six = "^1.16.0"
pillow = "^10.1.0"
python-consul = "^1.1.0"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
mypy = "^0.981"
pre-commit = "^2.20.0"
tox = "^3.25.1"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.2"
mkdocs-material = "^8.5.10"
mkdocstrings = {extras = ["python"], version = "^0.19.0"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py37']
preview = true

[tool.mypy]
files = ["preprocessing"]
disallow_untyped_defs = "True"
disallow_any_unimported = "True"
no_implicit_optional = "True"
check_untyped_defs = "True"
warn_return_any = "True"
warn_unused_ignores = "True"
show_error_codes = "True"

[tool.ruff]
target-version = "py37"
line-length = 120
fix = true
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]



